+++ src/index.js
@@ -1,1 +1,2 @@
-export { default } from "./checkable";
+export { default as checkable } from "./checkable";
+export { default as submitable } from "./submitable";